+++ packages/profile-sync-controller/src/controllers/user-storage/constants.ts
@@ -2,6 +2,7 @@ export const BACKUPANDSYNC_FEATURES = {
   main: 'main',
   accountSyncing: 'accountSyncing',
   contactSyncing: 'contactSyncing',
+  rampsSyncing: 'rampsSyncing',
 } as const;
 
 /**