+++ packages/ramps-controller/src/order-syncing/index.ts
@@ -1,0 +1,29 @@
+export {
+  USER_STORAGE_RAMPS_ORDERS_FEATURE,
+  USER_STORAGE_VERSION,
+  USER_STORAGE_VERSION_KEY,
+  TraceName,
+} from './constants.js';
+export type {
+  UserStorageRampsOrderEntry,
+  SyncRampsOrder,
+  OrderSyncingController,
+  OrderSyncingOptions,
+} from './types.js';
+export {
+  createOrderStorageKey,
+  isSyncableOrder,
+  mapRampsOrderToUserStorageEntry,
+  mapUserStorageEntryToRampsOrder,
+  stripPaymentDetailsForRemoteStorage,
+  stripSyncMetadata,
+  stripDeletedAt,
+  areOrdersEqual,
+} from './utils.js';
+export { canPerformOrderSyncing } from './sync-utils.js';
+export type { SyncOrdersWithUserStorageConfig } from './controller-integration.js';
+export {
+  syncOrdersWithUserStorage,
+  updateOrderInRemoteStorage,
+  deleteOrderInRemoteStorage,
+} from './controller-integration.js';